Make join game packet test more exhaustive
